+++ test/request_queue.js
@@ -32,6 +32,8 @@ describe('RequestQueue', () => {
             await queue.addRequest(new Apify.Request({ url: 'http://example.com/middle' }));
             await queue.addRequest(new Apify.Request({ url: 'http://example.com/last-but-first' }), { forefront: true });
 
+            expect(req1.id).to.be.a('string');
+            expect(req1.uniqueKey).to.eql('http://example.com/first');
             expect(info1).to.contain({
                 wasAlreadyPresent: false,
                 wasAlreadyHandled: false,
@@ -60,6 +62,14 @@ describe('RequestQueue', () => {
             expect(await queue.fetchNextRequest()).to.be.eql(null);
             expect(await queue.isEmpty()).to.be.eql(true);
             expect(await queue.isFinished()).to.be.eql(false);
+
+            // Test validations
+            await queue.markRequestHandled(new Apify.Request({ id: 'XXX', url: 'dummy' }))
+                .catch(err => expect(err.message).to.match(/Cannot mark request XXX as handled, because it is not in progress/));
+            await queue.reclaimRequest(new Apify.Request({ id: 'XXX', url: 'dummy' }))
+                .catch(err => expect(err.message).to.match(/Cannot reclaim request XXX, because it is not in progress/));
+            await queue.addRequest(new Apify.Request({ id: 'id-already-set', url: 'dummy' }))
+                .catch(err => expect(err.message).to.match(/Request already has the "id" field set so it cannot be added to the queue/));
 
             // Check that changes to Requests are persisted to Queue.
             request1.errorMessages = ['Hello'];
@@ -292,7 +302,33 @@ describe('RequestQueue', () => {
                     clientKey,
                 })
                 .returns(Promise.resolve({ requestId: 'a', wasAlreadyHandled: false, wasAlreadyPresent: false, request: requestA }));
-            await queue.addRequest(requestA);
+            const queueOperationInfo1 = await queue.addRequest(requestA);
+
+            expect(requestA.id).to.eql('a');
+            expect(requestA.uniqueKey).to.eql('http://example.com/a');
+            expect(queueOperationInfo1).to.include({
+                wasAlreadyPresent: false,
+                wasAlreadyHandled: false,
+                requestId: 'a',
+            });
+            expect(queueOperationInfo1.request).to.include({
+                id: 'a',
+            });
+            expect(queue.queueHeadDict.length()).to.be.eql(1);
+
+            // Try to add again the a request with the same URL
+            const copyOfRequest = { url: 'http://example.com/a' };
+            const queueOperationInfo2 = await queue.addRequest(copyOfRequest);
+            expect(copyOfRequest.id).to.eql('a');
+            expect(copyOfRequest.uniqueKey).to.eql('http://example.com/a');
+            expect(queueOperationInfo2).to.include({
+                wasAlreadyPresent: true,
+                wasAlreadyHandled: false,
+                requestId: 'a',
+            });
+            expect(queueOperationInfo2.request).to.include({
+                id: 'a',
+            });
             expect(queue.queueHeadDict.length()).to.be.eql(1);
 
             const requestB = new Request({ url: 'http://example.com/b' });
@@ -306,6 +342,7 @@ describe('RequestQueue', () => {
                 })
                 .returns(Promise.resolve({ requestId: 'b', wasAlreadyHandled: false, wasAlreadyPresent: false, request: requestB }));
             await queue.addRequest(requestB, { forefront: true });
+
             expect(queue.queueHeadDict.length()).to.be.eql(2);
             expect(queue.inProgressCount()).to.be.eql(0);
 
@@ -321,6 +358,14 @@ describe('RequestQueue', () => {
             expect(requestBFromQueue).to.be.eql(requestB);
             expect(queue.queueHeadDict.length()).to.be.eql(1);
             expect(queue.inProgressCount()).to.be.eql(1);
+
+            // Test validations
+            await queue.markRequestHandled(new Request({ id: 'XXX', url: 'dummy' }))
+                .catch(err => expect(err.message).to.match(/Cannot mark request XXX as handled, because it is not in progress/));
+            await queue.reclaimRequest(new Request({ id: 'XXX', url: 'dummy' }))
+                .catch(err => expect(err.message).to.match(/Cannot reclaim request XXX, because it is not in progress/));
+            await queue.addRequest(new Request({ id: 'id-already-set', url: 'dummy' }))
+                .catch(err => expect(err.message).to.match(/Request already has the "id" field set so it cannot be added to the queue/));
 
             // getRequest() returns null if object was not found.
             mock.expects('getRequest')
@@ -345,13 +390,12 @@ describe('RequestQueue', () => {
                 .returns(Promise.resolve({ requestId: requestB.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestB }));
             await queue.reclaimRequest(requestB, { forefront: true });
             expect(queue.queueHeadDict.length()).to.be.eql(1);
-            expect(queue.inProgressCount()).to.be.eql(0);
+            expect(queue.inProgressCount()).to.be.eql(1);
             await Apify.utils.sleep(STORAGE_CONSISTENCY_DELAY_MILLIS + 10);
             expect(queue.queueHeadDict.length()).to.be.eql(2);
             expect(queue.inProgressCount()).to.be.eql(0);
 
             // Fetch again.
-            expect(queue.queueHeadDict.length()).to.be.eql(1);
             mock.expects('getRequest')
                 .once()
                 .withArgs({
@@ -376,6 +420,9 @@ describe('RequestQueue', () => {
             await queue.markRequestHandled(requestB);
             expect(queue.queueHeadDict.length()).to.be.eql(1);
             expect(queue.inProgressCount()).to.be.eql(0);
+
+            // Emulate there are no cached items in queue
+            queue.queueHeadDict.clear();
 
             // Query queue head.
             mock.expects('getHead')
@@ -566,6 +613,7 @@ describe('RequestQueue', () => {
                     request: requestA,
                 }));
             await queue.addRequest(requestA, { forefront: true });
+            expect(queue.queueHeadDict.length()).to.be.eql(1);
 
             // Try to get requestA which is not available yet.
             mock.expects('getRequest')
@@ -577,7 +625,10 @@ describe('RequestQueue', () => {
                 .returns(Promise.resolve(null));
             expect(await queue.fetchNextRequest()).to.be.eql(null);
 
-            // Should try it once again.
+            // Give queue time to mark request 'a' as not in progress
+            await Apify.utils.sleep(STORAGE_CONSISTENCY_DELAY_MILLIS + 10);
+
+            // Should try it once again (the queue head is queried again)
             mock.expects('getRequest')
                 .once()
                 .withArgs({
@@ -585,6 +636,18 @@ describe('RequestQueue', () => {
                     requestId: 'a',
                 })
                 .returns(Promise.resolve(requestA));
+            mock.expects('getHead')
+                .once()
+                .withArgs({
+                    queueId: 'some-id',
+                    limit: QUERY_HEAD_MIN_LENGTH,
+                    clientKey,
+                })
+                .returns(Promise.resolve({
+                    items: [
+                        { id: 'a', uniqueKey: 'aaa' },
+                    ],
+                }));
             expect(await queue.fetchNextRequest()).to.be.eql(requestA);
 
             mock.verify();
@@ -771,7 +834,7 @@ describe('RequestQueue', () => {
                 .returns(Promise.resolve({ requestId: requestA.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestA }));
             await queue.reclaimRequest(requestA, { forefront: true });
             expect(queue.queueHeadDict.length()).to.be.eql(0);
-            expect(queue.inProgressCount()).to.be.eql(0);
+            expect(queue.inProgressCount()).to.be.eql(1);
             expect(queue.assumedTotalCount).to.be.eql(2);
             expect(queue.assumedHandledCount).to.be.eql(1);
             await Apify.utils.sleep(STORAGE_CONSISTENCY_DELAY_MILLIS + 10);